Added face turn functionality to left, right, and top faces
There is some bug with the meshes that needs to be investigated. Process to see the bug:
Isometric view.
Turn top face right.
Turn right face down.

The red color on top turns down to the front correctly. But the side colors don't move properly (maybe rotate the opposite direction?).